Added new service capability-onboarding-service
