+++ libs/astra-models/pyproject.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "astra-models"
+version = "0.1.0"
+requires-python = ">=3.11"
+description = "Shared domain models for ASTRA services"
+dependencies = ["pydantic>=2.8"]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["astra_models*"]
